resources/requirements.txt
Pillow
numpy
piexif
pyzipper
pycryptodome